# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 英語教師教室管理與學習輔助系統 API
  version: 1.0.0-m0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    華語地區英文教師教室管理系統的 REST API 規格（M0 定稿）。

    - 路徑統一前綴 `/v1`；資源名詞複數；最多三層資源。
    - 一律 JSON、欄位 camelCase、日期 ISO 8601。
    - 錯誤回應統一格式 `{ error: { code, message, details } }`，錯誤碼格式 `模組_錯誤類型`。
    - 認證：`Authorization: Bearer <accessToken>`（JWT，15 分鐘）；refresh token 走 httpOnly cookie。
servers:
  - url: https://your-teacher-server.example
    description: 教師自架實例（替換為實際網域）
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      operationId: registerTeacher
      summary: 教師註冊
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterTeacherRequest" }
      responses:
        "201":
          description: 註冊成功，回傳 token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Email 已被註冊
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: AUTH_EMAIL_ALREADY_USED
                  message: 此 email 已註冊
                  details: { email: teacher@example.com }
  /v1/auth/register/student:
    post:
      tags: [Auth]
      operationId: registerStudent
      summary: 學生以班級邀請碼自助註冊
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterStudentRequest" }
      responses:
        "201":
          description: 註冊成功並自動加入邀請碼對應班級
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: 邀請碼過期或無效
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: AUTH_INVITE_CODE_INVALID
                  message: 邀請碼無效或已過期
                  details: { inviteCode: AB3D-7K9M }
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 登入（教師／學生／助教共用）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 登入成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: 帳號或密碼錯誤
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: AUTH_INVALID_CREDENTIALS
                  message: 帳號或密碼錯誤
                  details: {}
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: 以 refresh token 換發新 access token（rotation）
      security: []
      parameters:
        - in: cookie
          name: refreshToken
          required: true
          schema: { type: string }
          description: 登入時以 httpOnly cookie 下發
      responses:
        "200":
          description: 換發成功（舊 refresh token 作廢、下發新 cookie）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: refresh token 無效 / 過期 / 已被撤銷（偵測重用時整條 family 撤銷）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: AUTH_REFRESH_TOKEN_INVALID
                  message: refresh token 無效，請重新登入
                  details: {}
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 登出（撤銷目前 refresh token）
      responses:
        "204": { description: 已登出並清除 cookie }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: 參數錯誤（400）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: COMMON_VALIDATION_FAILED, message: 請求參數驗證失敗, details: { field: page, reason: 須為正整數 } }
    Unauthorized:
      description: 未認證（401）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: AUTH_TOKEN_MISSING, message: 未提供存取憑證, details: {} }
...
    NotFound:
      description: 資源不存在（404）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: CLASS_NOT_FOUND, message: 找不到該班級, details: {} }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 機器可讀錯誤碼，格式 `模組_錯誤類型`
              example: CLASS_NOT_FOUND
            message: { type: string, description: 人類可讀訊息（繁體中文） }
            details:
              type: object
              additionalProperties: true
              description: 上下文（資源 ID 等），不含敏感資訊
...
    AuthTokens:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken: { type: string, description: JWT，15 分鐘效期 }
        expiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/UserInfo" }
    UserInfo:
      type: object
      required: [id, role, name]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [TEACHER, STUDENT, ASSISTANT] }
        name: { type: string }
        email: { type: string, format: email }
        username: { type: string }
    RegisterTeacherRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string, maxLength: 60 }
    RegisterStudentRequest:
      type: object
      required: [inviteCode, password, name]
      properties:
        inviteCode: { type: string, example: AB3D-7K9M }
        username: { type: string, description: 與 email 二選一 }
        email: { type: string, format: email, description: 與 username 二選一 }
        password: { type: string, minLength: 8 }
        name: { type: string, maxLength: 60 }
    LoginRequest:
      type: object
      required: [password]
      properties:
        email: { type: string, format: email }
        username: { type: string }
        password: { type: string }
      description: email 或 username 擇一
...